The admin logic has been moved to users dir, create admin methods in user repositories and services
