--- src/admin/routers.py
@@ -1,3 +1,0 @@
-from fastapi import APIRouter
-
-router = APIRouter(tags=["admin"], prefix="/admin")
