+++ src/users/repositories.py
@@ -1,5 +1,5 @@
 import random
-from typing import Optional
+from typing import Optional, List
 
 from sqlalchemy import insert, select, delete, update
 
@@ -77,12 +77,47 @@ class UserRepository:
             user = result.scalars().first()
         return user
 
+    async def get_all_users(self) -> List[User]:
+        async with async_session() as session:
+            query = select(User)
+            result = await session.execute(query)
+            users = result.scalars().all()
+
+            return users
+
     async def get_user_by_short_name(self, short_name: str) -> Optional[User]:
         async with async_session() as session:
             query = select(User).where(User.short_name == short_name)
             result = await session.execute(query)
             user = result.scalars().first()
         return user
+
+    async def change_admin_status(self, user: User) -> User:
+        async with async_session() as session:
+            stmt = update(User).where(User.id == user.id).values(is_admin=False if user.is_admin else True)
+            await session.execute(stmt)
+            await session.commit()
+
+            user: User = await self.get_user_by_id(user.id)
+            return user
+
+    async def change_verified_status(self, user: User) -> User:
+        async with async_session() as session:
+            stmt = update(User).where(User.id == user.id).values(is_verified=False if user.is_verified else True)
+            await session.execute(stmt)
+            await session.commit()
+
+            user: User = await self.get_user_by_id(user.id)
+            return user
+
+    async def change_active_status(self, user: User) -> User:
+        async with async_session() as session:
+            stmt = update(User).where(User.id == user.id).values(is_active=False if user.is_active else True)
+            await session.execute(stmt)
+            await session.commit()
+
+            user: User = await self.get_user_by_id(user.id)
+            return user
 
     async def get_user_by_id(self, user_id: int) -> Optional[User]:
         async with async_session() as session:
